feat: add terraform pool group package
