+++ libs/terraform/README.md
@@ -1,0 +1,3 @@
+TODO to clean:
+
+- [ ] make container api docker tag configurable
