+++ libs/terraform/pool-group/src/lib/configs/index.ts
@@ -1,0 +1,1 @@
+export * from './pool-group.config';
