+++ libs/terraform/pool-group/src/lib/configs/pool-group.config.ts
@@ -1,0 +1,9 @@
+import { PoolGroupConfig } from '../interfaces';
+
+export const poolGroupDefaultConfig: Pick<
+  PoolGroupConfig,
+  'subDomain' | 'reverseProxyDesiredCount'
+> = {
+  subDomain: 'www',
+  reverseProxyDesiredCount: 1,
+};
