+++ libs/terraform/pool-group/src/lib/containers/backend-container.ts
@@ -1,4 +1,5 @@
-import { Container, ContainerConfig, ContainerDefinition } from './container';
+import { ContainerConfig, ContainerDefinition } from '../interfaces';
+import { Container } from './container';
 
 /**
  * This container is meant to be used by another one, as its backend
