+++ libs/terraform/pool-group/src/lib/deployment/deployment-pipeline.component.ts
@@ -2,8 +2,10 @@ import { Codepipeline, CodepipelineStage } from '@cdktf/provider-aws';
 import { Token } from 'cdktf';
 import { ConstructOptions } from 'constructs';
 
-import { AwsComponent } from '../../../abstracts/aws.component';
-import { AwsProviderConstruct } from '../../../abstracts/aws.interface';
+import {
+  AwsComponent,
+  AwsProviderConstruct,
+} from '@tractr/terraform-aws-component';
 
 export interface DeploymentTrigger {
   repositoryName: string;
