+++ libs/terraform/pool-group/src/lib/deployment/deployment-role.component.ts
@@ -2,8 +2,10 @@ import { IamRole, IamRoleInlinePolicy } from '@cdktf/provider-aws';
 import { Token } from 'cdktf';
 import { ConstructOptions } from 'constructs';
 
-import { AwsComponent } from '../../../abstracts/aws.component';
-import { AwsProviderConstruct } from '../../../abstracts/aws.interface';
+import {
+  AwsComponent,
+  AwsProviderConstruct,
+} from '@tractr/terraform-aws-component';
 
 export interface DeploymentRoleComponentConfig extends ConstructOptions {
   storeS3Arn: string;
