+++ libs/terraform/pool-group/src/lib/deployment/deployment-store.component.ts
@@ -2,8 +2,10 @@ import { S3Bucket } from '@cdktf/provider-aws';
 import { Token } from 'cdktf';
 import { ConstructOptions } from 'constructs';
 
-import { AwsComponent } from '../../../abstracts/aws.component';
-import { AwsProviderConstruct } from '../../../abstracts/aws.interface';
+import {
+  AwsComponent,
+  AwsProviderConstruct,
+} from '@tractr/terraform-aws-component';
 
 export class DeploymentStoreComponent extends AwsComponent<ConstructOptions> {
   protected readonly s3Bucket: S3Bucket;
