+++ libs/terraform/pool-group/src/lib/deployment/trigger/deployment-trigger-role.component.ts
@@ -6,8 +6,10 @@ import {
 import { Token } from 'cdktf';
 import { ConstructOptions } from 'constructs';
 
-import { AwsComponent } from '../../../../abstracts/aws.component';
-import { AwsProviderConstruct } from '../../../../abstracts/aws.interface';
+import {
+  AwsComponent,
+  AwsProviderConstruct,
+} from '@tractr/terraform-aws-component';
 
 export interface DeploymentTriggerRoleComponentConfig extends ConstructOptions {
   codepipelineArn: string;
