+++ libs/terraform/pool-group/src/lib/interfaces/index.ts
@@ -1,0 +1,2 @@
+export * from './container.interface';
+export * from './pool-group.interface';
