+++ libs/terraform/pool-group/src/lib/logs.component.ts
@@ -1,8 +1,10 @@
 import { CloudwatchLogGroup, CloudwatchLogStream } from '@cdktf/provider-aws';
 import { ConstructOptions } from 'constructs';
 
-import { AwsComponent } from '../../abstracts/aws.component';
-import { AwsProviderConstruct } from '../../abstracts/aws.interface';
+import {
+  AwsComponent,
+  AwsProviderConstruct,
+} from '@tractr/terraform-aws-component';
 
 export class LogsComponent extends AwsComponent<ConstructOptions> {
   protected readonly cloudwatchLogGroup: CloudwatchLogGroup;
