+++ libs/terraform/pool-group/src/lib/private-dns/private-dns.component.ts
@@ -2,8 +2,10 @@ import { ServiceDiscoveryPrivateDnsNamespace } from '@cdktf/provider-aws';
 import { Token } from 'cdktf';
 import { ConstructOptions } from 'constructs';
 
-import { AwsComponent } from '../../../abstracts/aws.component';
-import { AwsProviderConstruct } from '../../../abstracts/aws.interface';
+import {
+  AwsComponent,
+  AwsProviderConstruct,
+} from '@tractr/terraform-aws-component';
 
 export interface PrivateDnsComponentConfig extends ConstructOptions {
   vpcId: string;
