+++ libs/terraform/pool-group/src/lib/reverse-proxy/reverse-proxy-task-role.component.ts
@@ -2,8 +2,10 @@ import { IamRole, IamRoleInlinePolicy } from '@cdktf/provider-aws';
 import { Token } from 'cdktf';
 import { ConstructOptions } from 'constructs';
 
-import { AwsComponent } from '../../../abstracts/aws.component';
-import { AwsProviderConstruct } from '../../../abstracts/aws.interface';
+import {
+  AwsComponent,
+  AwsProviderConstruct,
+} from '@tractr/terraform-aws-component';
 
 export class ReverseProxyTaskRoleComponent extends AwsComponent<ConstructOptions> {
   protected readonly iamRole: IamRole;
