+++ libs/terraform/pool-group/src/lib/secrets/secrets.component.ts
@@ -2,8 +2,10 @@ import { SecretsmanagerSecret } from '@cdktf/provider-aws';
 import { Token } from 'cdktf';
 import { ConstructOptions } from 'constructs';
 
-import { AwsComponent } from '../../../abstracts/aws.component';
-import { AwsProviderConstruct } from '../../../abstracts/aws.interface';
+import {
+  AwsComponent,
+  AwsProviderConstruct,
+} from '@tractr/terraform-aws-component';
 
 export class SecretsComponent extends AwsComponent<ConstructOptions> {
   protected readonly secretsmanagerSecret: SecretsmanagerSecret;
