+++ package.json
@@ -82,6 +82,7 @@
     "fs-extra": "^10.0.0",
     "graphql": "^15.5.1",
     "jest-mock-extended": "^1.0.18",
+    "md5": "^2.3.0",
     "mime-types": "^2.1.32",
     "minio": "^7.0.18",
     "morgan": "^1.10.0",
@@ -138,6 +139,7 @@
     "@types/fs-extra": "^9.0.12",
     "@types/jest": "26.0.24",
     "@types/jsonwebtoken": "^8.5.4",
+    "@types/md5": "^2.3.1",
     "@types/mime-types": "^2.1.0",
     "@types/minio": "^7.0.9",
     "@types/morgan": "^1.9.3",
